Begin removing configure_submitty and converting to generate_configs and validate_configs
