+++ .setup/data/configs/authentication.json
@@ -1,0 +1,12 @@
+{
+    "authentication_method": "PamAuthentication",
+    "ldap_options": {
+        "url": "",
+        "uid": "",
+        "bind_dn": ""
+    },
+    "saml_options": {
+        "name": "",
+        "username_attribute": ""
+    }
+}
