+++ .setup/data/configs/database.json
@@ -1,0 +1,10 @@
+{
+    "authentication_method": "PamAuthentication",
+    "database_host": "localhost",
+    "database_port": 5432,
+    "database_user": "submitty_dbuser",
+    "database_user_password": "submitty_dbuser_password",
+    "database_course_user": "submitty_course_dbuser",
+    "database_course_user_password": "submitty_course_dbuser",
+    "debugging_enabled": false
+}
