+++ .setup/data/configs/submitty_admin.json
@@ -1,0 +1,3 @@
+{
+    "submitty_admin_username": "submitty-admin"
+}
