+++ .setup/data/configs/worker/submitty.json
@@ -1,0 +1,6 @@
+config['submitty_install_dir'] = SUBMITTY_INSTALL_DIR
+config['submitty_repository'] = SUBMITTY_REPOSITORY
+config['submitty_data_dir'] = SUBMITTY_DATA_DIR
+config['autograding_log_path'] = AUTOGRADING_LOG_PATH
+config['site_log_path'] = TAGRADING_LOG_PATH
+"worker": true
